Move notebook files
