+++ src/notebook/default-toolbar.ts
