+++ src/notebook/index.ts
@@ -1,5 +1,12 @@
 // Copyright (c) Jupyter Development Team.
 // Distributed under the terms of the Modified BSD License.
 
-export * from './notebook/index';
+export * from './actions';
+export * from './default-toolbar';
+export * from './model';
+export * from './modelfactory';
+export * from './panel';
 export * from './tracker';
+export * from './trust';
+export * from './widget';
+export * from './widgetfactory';
