--- src/notebook/notebook/index.ts
@@ -1,12 +1,0 @@
-// Copyright (c) Jupyter Development Team.
-// Distributed under the terms of the Modified BSD License.
-
-export * from './actions';
-export * from './default-toolbar';
-export * from './model';
-export * from './modelfactory';
-export * from './panel';
-export * from './trust';
-export * from './widget';
-export * from './widgetfactory';
-
