+++ src/notebook/panel.ts
