+++ src/notebook/widget.ts
