+++ src/notebook/widgetfactory.ts
